@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net7.0-windows</TargetFramework>
     <UseWindowsForms>true</UseWindowsForms>
     <Copyright>Roger Briggen</Copyright>
     <PackageProjectUrl>https://github.com/rogerbriggen/cloudBackup</PackageProjectUrl>
@@ -24,7 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\cloudBackupCore\cloudBackupCore.csproj" />
+    <ProjectReference Include="..\cloudBackupCore\CloudBackupCore.csproj" />
   </ItemGroup>
 
   <ItemGroup>
